+++ django-data/image/accounts/tests/test_view_password_reset.py
@@ -96,9 +96,9 @@ class PasswordResetConfirmTests(TestCase):
         '''
         create a valid password reset token
         based on how django creates the token internally:
-        https://github.com/django/django/blob/1.11.5/django/contrib/auth/forms.py#L280
+        https://github.com/django/django/blob/612c2d166c9048c59ee241ce4ba89858aa65665d/django/contrib/auth/forms.py#L287
         '''
-        self.uid = urlsafe_base64_encode(force_bytes(user.pk)).decode()
+        self.uid = urlsafe_base64_encode(force_bytes(user.pk))
         self.token = default_token_generator.make_token(user)
 
         url = reverse(
@@ -135,7 +135,7 @@ class InvalidPasswordResetConfirmTests(TestCase):
     def setUp(self):
         user = User.objects.create_user(
             username='john', email='john@doe.com', password='123abcdef')
-        uid = urlsafe_base64_encode(force_bytes(user.pk)).decode()
+        uid = urlsafe_base64_encode(force_bytes(user.pk))
         token = default_token_generator.make_token(user)
 
         '''